refactor: Extract login logic into a custom hook (useLogin)
- Moved login logic from LoginPopUp component to a new custom hook named useLogin.
- Separated state management and API interaction from the UI component.
- Improved code maintainability and reusability by isolating business logic from UI rendering.
- Implemented global login popup in Card and Related Product components.
